hotspot visualisation development
