+++ js/actions/nglLoadActions.js
@@ -1,9 +1,12 @@
 /**
  * Created by abradley on 03/03/2018.
  */
-import {LOAD_OBJECT, LOAD_OBJECT_SUCCESS, LOAD_OBJECT_FAILURE, OBJECT_LOADING,
+import {
+    LOAD_OBJECT, LOAD_OBJECT_SUCCESS, LOAD_OBJECT_FAILURE, OBJECT_LOADING,
     DELETE_OBJECT, DELETE_OBJECT_FAILURE, DELETE_OBJECT_SUCCESS,
-    DELETE_OBJECT_TYPE, SET_ORIENTATION, SET_NGL_ORIENTATION, SET_UUID, SET_LOADING_STATE} from './actonTypes'
+    DELETE_OBJECT_TYPE, SET_ORIENTATION, SET_NGL_ORIENTATION, SET_UUID, SET_LOADING_STATE,
+    SET_STAGE_COLOR
+} from './actonTypes'
 
 
 export const loadObject = function (group) {
@@ -112,5 +115,13 @@ export const setLoadingState = function (bool) {
     return {
         type: SET_LOADING_STATE,
         loadingState: bool
+    };
+}
+
+export const setStageColor = function (stageColor) {
+    console.log("ACTIONS: " + stageColor);
+    return {
+        type: SET_STAGE_COLOR,
+        stageColor: stageColor
     };
 }
